Refactor spell functions into spells.py
