+++ lament_mod/lament.py
@@ -4,6 +4,7 @@ from flask import request, send_file, render_template, flash, url_for, redirect,
 
 import lament_mod.character as character
 import lament_mod.tools as tools
+import lament_mod.spells as spells
 from fdfgen import forge_fdf
 from PyPDF2 import PdfFileMerger, PdfFileReader
 import subprocess
@@ -93,7 +94,7 @@ def lament_pdf():
         # If the character has spells, create a PDF spell sheet and fill
         # it with spells and spell info
         if PC.pcClass in ['Cleric', 'Magic-User', 'Elf']:
-            tools.create_spellsheet_pdf(
+            spells.create_spellsheet_pdf(
                 PC.details,
                 PC.name,
                 filename=None,
